Add files via upload
A stored procedure is a piece of code that you can save and reuse in your database. In this case, we'll create a stored procedure that updates the `updated_at` column whenever new data is inserted or updated. We'll then create a trigger that calls this stored procedure every time the table is updated.